makefile !

+++ waf_tools/any_lapack.py
@@ -255,7 +255,7 @@ void dsymm(const char *side, const char *uplo, const int *m, const int *n,
            const double *alpha, const double *a, const int *lda, const double *b, const int *ldb,
            const double *beta, double *c, const int *ldc);
 double ddot(int* N,double *DX, int* INCX,double *DY,int* INCY);
-void dpotrs(char* UPLO,int * N,int * NRHS,double* A,int* LDA,double* B,int* LDB,double* INFO );
+void dpotrs(char* UPLO,int * N,int * NRHS,double* A,int* LDA,double* B,int* LDB,int* INFO );
            
          
 """
